+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### dev-master
 
-* #3: escape(null) returns empty string not type-error
+* GH-3: escape(null) returns empty string not type-error
+* GH-2: some more formats for `HTML::select()` options
 
 ### 1.0.0 (07.07.2019)